Time: 21 ms (66.29%), Space: 22.4 MB (92.34%) - LeetHub

+++ 637-average-of-levels-in-binary-tree/637-average-of-levels-in-binary-tree.cpp
@@ -23,7 +23,7 @@ public:
         while(!q.empty()){
             int n=q.size();
             double avgg=0;
-            int ns=n;
+            double ns=n;
             while(n--){
                 TreeNode *temp=q.front();
                 q.pop();
